i forgot i fucked up something

+++ src/include/string.h
@@ -5,7 +5,7 @@
 char *itoa(int value, char *buffer, int base);
 int atoi(const char* str);
 
-void *memset(void *buffer, int value, size_t size);
+void *memset(void *buffer, uint8_t value, size_t size);
 void *memcpy(void *dest, const void *src, size_t size);
 int memcmp(const void *s1, const void *s2, size_t size);
 void *memmove(void *dest, const void *src, size_t size);
